chore: Remove bundle naming
