+++ .env
@@ -117,8 +117,6 @@ E2E_TEST_ACCOUNT_BASE_PASSWORD=test-password
 # Monitor Premium features
 # Link to start user on the subscription process. PREMIUM_ENABLED must be set to `true`.
 SUBSCRIPTION_BILLING_AMOUNT_MONTHLY_US=42.42
-SUBSCRIPTION_BILLING_AMOUNT_BUNDLE_INDIVIDUAL_MONTHLY_US=424
-SUBSCRIPTION_BILLING_AMOUNT_BUNDLE_MONTHLY_US=42
 
 # SubPlat 2.0 URL, product and plan IDs, used for Plus subscriptions:
 FXA_SUBSCRIPTIONS_URL=https://accounts.stage.mozaws.net/subscriptions
