+++ .env.local.example
@@ -43,8 +43,6 @@ EMAIL_FROM="Mozilla Monitor Stage <breach-alerts@stage.mozaws.net>"
 # Whether GA4 sends data or not. NOTE: must be set in build environment.
 NEXT_PUBLIC_GA4_DEBUG_MODE=true
 SUBSCRIPTION_BILLING_AMOUNT_MONTHLY_US=42.42
-SUBSCRIPTION_BILLING_AMOUNT_BUNDLE_INDIVIDUAL_MONTHLY_US=424
-SUBSCRIPTION_BILLING_AMOUNT_BUNDLE_MONTHLY_US=42
 
 # HIBP setup
 HIBP_KANON_API_ROOT=https://enterprise.stage-api.haveibeenpwned.com
