+++ .github/workflows/functional_tests_pr.yml
@@ -83,6 +83,7 @@ jobs:
           ONEREP_API_BASE: "http://localhost:6060/api/mock/onerep/"
           ONEREP_API_KEY: ${{ secrets.ONEREP_API_KEY }}
           PREMIUM_PLAN_ID_MONTHLY_US: ${{ secrets.STAGE_PREMIUM_PLAN_ID_MONTHLY_US }}
+          PREMIUM_PLAN_ID_YEARLY_US: ${{ secrets.STAGE_PREMIUM_PLAN_ID_YEARLY_US }}
           PREMIUM_PRODUCT_ID: ${{ secrets.STAGE_PREMIUM_PRODUCT_ID }}
           REDIS_URL: "redis://redis.mock"
       - uses: actions/upload-artifact@v4
