+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/settings/stories/SettingsStoryWrapper.tsx
@@ -100,9 +100,7 @@ export const SettingsWrapper = (props: SettingsWrapperProps) => {
             fxaSettingsUrl=""
             fxaSubscriptionsUrl=""
             monthlySubscriptionUrl=""
-            yearlySubscriptionUrl=""
             subscriptionBillingAmount={{
-              yearly: 13.37,
               monthly: 42.42,
             }}
             enabledFeatureFlags={props.enabledFeatureFlags}
