+++ src/app/(proper_react)/(redesign)/(public)/LandingViewRedesign/components/PricingPlanList.tsx
@@ -12,25 +12,36 @@ import { modifyAttributionsForUrlSearchParams } from "../../../../../functions/u
 import { TelemetryButton } from "../../../../../components/client/TelemetryButton";
 import { VisuallyHidden } from "../../../../../components/server/VisuallyHidden";
 import {
+  CONST_MAX_NUM_ADDRESSES,
+  CONST_MAX_NUM_ADDRESSES_PLUS,
   CONST_ONEREP_DATA_BROKER_COUNT,
   CONST_URL_MONITOR_LANDING_PAGE_ID,
 } from "../../../../../../constants";
 import styles from "./PricingPlanList.module.scss";
-import { CheckIcon } from "../../../../../components/server/Icons";
+import { CheckIcon, PlusIcon } from "../../../../../components/server/Icons";
 import { ScanLimit } from "../../ScanLimit";
 import { FreeScanCta } from "../../FreeScanCta";
 import { ExperimentData } from "../../../../../../telemetry/generated/nimbus/experiments";
 import { FeatureFlagName } from "../../../../../../db/tables/featureFlags";
+import { usePathname } from "next/navigation";
+import { BundleBillingAmount } from "../../../../../functions/server/getPremiumSubscriptionInfo";
+
+export type ProductBundleUrl = Record<"relay" | "vpn", string>;
+
+export type PremiumSubscriptionUrl = {
+  monthly: string;
+};
+
+export type SubscriptionBillingAmount = {
+  monthly: number;
+  bundle: BundleBillingAmount;
+};
 
 export type Props = {
-  "aria-labelledby": string;
-  premiumSubscriptionUrl: {
-    monthly: string;
-  };
-  subscriptionBillingAmount: {
-    monthly: number;
-  };
+  premiumSubscriptionUrl: PremiumSubscriptionUrl;
+  subscriptionBillingAmount: SubscriptionBillingAmount;
   enabledFeatureFlags: FeatureFlagName[];
+  hideFreeCard?: boolean;
 };
 
 type ScanLimitProp = {
@@ -42,44 +53,54 @@ type ScanLimitProp = {
 type PricingPlanData = {
   type: "free" | "plus";
   title: string;
-  subtitle: string;
+  subtitle: ReactNode | string;
   features: ReactNode[];
   cta: ReactNode;
-  label?: string;
+  label?: ReactNode;
 };
 
 export const PricingPlanList = (props: Props & ScanLimitProp) => {
   const l10n = useL10n();
+  const pathname = usePathname();
+  const isLandingPage = pathname === "/";
 
   // The cookie `attributionsLastTouch` is set in the component `PageLoadEvent`
   // to help with attributions.
   const [cookies] = useCookies(["attributionsLastTouch"]);
-  const newSearchParam = modifyAttributionsForUrlSearchParams(
+  const landingPageSearchParams = {
+    utm_source: "monitor-product",
+    utm_medium: "monitor",
+    utm_campaign: "landing-page-pricing-grid",
+    utm_content: "pricing-grid-us",
+  };
+  const subscriptionPlansPageSearchParams = {
+    utm_source: "monitor-product",
+    utm_medium: "monitor",
+    utm_campaign: "in-product-pricing-grid",
+    utm_content: "get-started-us",
+  };
+  const initSearchParams = {
+    entrypoint: CONST_URL_MONITOR_LANDING_PAGE_ID,
+    form_type: "button",
+    data_cta_position: "pricing",
+  };
+  const searchParams = isLandingPage
+    ? landingPageSearchParams
+    : subscriptionPlansPageSearchParams;
+  const newSearchParamPlus = modifyAttributionsForUrlSearchParams(
     new URLSearchParams(cookies.attributionsLastTouch),
-    {
-      entrypoint: CONST_URL_MONITOR_LANDING_PAGE_ID,
-      form_type: "button",
-      data_cta_position: "pricing",
-    },
-    {
-      utm_source: "product",
-      utm_medium: "monitor",
-      utm_campaign: "pricing",
-    },
+    initSearchParams,
+    searchParams,
   );
   // SubPlat2 subscription links already have the UTM parameter `?plan` appended.
-  const additionalSubplatParamsString = `${props.enabledFeatureFlags.includes("SubPlat3") ? "?" : "&"}${newSearchParam.toString()}`;
+  const additionalSubplatParamsStringPlus = `${props.enabledFeatureFlags.includes("SubPlat3") ? "?" : "&"}${newSearchParamPlus.toString()}`;
 
-  const roundedPriceFormatter = new Intl.NumberFormat(getLocale(l10n), {
-    style: "currency",
-    currency: "USD",
-    currencyDisplay: "narrowSymbol",
-    maximumFractionDigits: 0,
-  });
   const priceFormatter = new Intl.NumberFormat(getLocale(l10n), {
     style: "currency",
     currency: "USD",
     currencyDisplay: "narrowSymbol",
+    minimumFractionDigits: 0,
+    maximumFractionDigits: 2,
   });
   const { monthly: monthlyPriceMonthlyBilling } =
     props.subscriptionBillingAmount;
@@ -87,44 +108,108 @@ export const PricingPlanList = (props: Props & ScanLimitProp) => {
   const pricingPlanData: PricingPlanData[] = [
     {
       type: "plus",
-      label: l10n.getString("landing-redesign-pricing-plans-card-plus-label"),
       title: l10n.getString("landing-redesign-pricing-plans-card-plus-title"),
       subtitle: l10n.getString(
-        "landing-redesign-pricing-plans-card-plus-subtitle",
+        "landing-redesign-pricing-plans-card-plus-with-bundle-subtitle",
       ),
       features: [
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-plus-feature-item-one",
-          {
-            vars: {
-              data_broker_sites_total_num: CONST_ONEREP_DATA_BROKER_COUNT,
-            },
-            elems: { b: <b /> },
-          },
-        ),
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-plus-feature-item-two",
-          {
-            elems: { b: <b /> },
-          },
-        ),
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-plus-feature-item-three",
-          {
-            elems: { b: <b /> },
-          },
-        ),
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-plus-feature-item-four",
-          {
-            elems: { b: <b /> },
-          },
-        ),
+        <>
+          <span className={styles.plusNote}>
+            <PlusIcon
+              alt={l10n.getString(
+                "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+              )}
+            />
+            <b>
+              {l10n.getString(
+                "landing-redesign-pricing-plans-card-plus-with-bundle-included-features-label",
+              )}
+            </b>
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {/* c8 ignore next 4 */}
+            {props.enabledFeatureFlags.includes("MaskDataBrokerCount")
+              ? l10n.getFragment(
+                  "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-one-masked",
+                  { elems: { b: <b /> } },
+                )
+              : l10n.getFragment(
+                  "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-one",
+                  {
+                    vars: {
+                      data_broker_sites_total_num:
+                        CONST_ONEREP_DATA_BROKER_COUNT,
+                    },
+                    elems: { b: <b /> },
+                  },
+                )}
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {l10n.getFragment(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-two",
+              {
+                elems: { b: <b /> },
+              },
+            )}
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {l10n.getFragment(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-three",
+              {
+                elems: { b: <b /> },
+              },
+            )}
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {l10n.getFragment(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-four",
+              {
+                elems: { b: <b /> },
+                vars: {
+                  max_email_addresses: CONST_MAX_NUM_ADDRESSES_PLUS,
+                },
+              },
+            )}
+          </span>
+        </>,
       ],
       cta: (
         <>
-          <p aria-live="polite">
-            <strong id="pricingPlan">
+          <p aria-live="polite" id="pricingPlan">
+            <span className={styles.pricingCardSavings}>
+              {l10n.getString(
+                "landing-redesign-pricing-plans-card-cta-monthly-billing-label",
+              )}
+            </span>
+            <strong>
               {l10n.getString(
                 "landing-redesign-pricing-plans-card-plus-cta-monthly",
                 {
@@ -136,10 +221,11 @@ export const PricingPlanList = (props: Props & ScanLimitProp) => {
             </strong>
           </p>
           <TelemetryButton
-            aria-describedby="pricingPlan pricingPlansReassurancePlus"
+            aria-describedby="pricingPlan"
             disabled={props.scanLimitReached}
-            variant="primary"
-            href={`${props.premiumSubscriptionUrl["monthly"]}${additionalSubplatParamsString}`}
+            variant="secondary"
+            theme="blue"
+            href={`${props.premiumSubscriptionUrl["monthly"]}${additionalSubplatParamsStringPlus}`}
             event={{
               module: "upgradeIntent",
               name: "click",
@@ -148,9 +234,7 @@ export const PricingPlanList = (props: Props & ScanLimitProp) => {
               },
             }}
           >
-            {l10n.getString(
-              "landing-redesign-pricing-plans-card-plus-cta-label",
-            )}
+            {l10n.getString("landing-redesign-pricing-plans-card-cta-label")}
           </TelemetryButton>
         </>
       ),
@@ -159,51 +243,111 @@ export const PricingPlanList = (props: Props & ScanLimitProp) => {
       type: "free",
       title: l10n.getString("landing-redesign-pricing-plans-card-free-title"),
       subtitle: l10n.getString(
-        "landing-redesign-pricing-plans-card-free-subtitle",
+        "landing-redesign-pricing-plans-card-free-with-bundle-subtitle",
       ),
       features: [
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-free-feature-item-one",
-          {
-            vars: {
-              data_broker_sites_total_num: CONST_ONEREP_DATA_BROKER_COUNT,
-            },
-            elems: { b: <b /> },
-          },
-        ),
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-free-feature-item-two",
-          {
-            elems: { b: <b /> },
-          },
-        ),
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-free-feature-item-three",
-          {
-            elems: { b: <b /> },
-          },
-        ),
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-free-feature-item-four",
-          {
-            elems: { b: <b /> },
-          },
-        ),
-        l10n.getFragment(
-          "landing-redesign-pricing-plans-card-free-feature-item-five",
-          {
-            elems: { b: <b /> },
-          },
-        ),
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {
+              /* c8 ignore start */
+              props.enabledFeatureFlags.includes("MaskDataBrokerCount")
+                ? l10n.getFragment(
+                    "landing-redesign-pricing-plans-card-free-with-bundle-feature-item-one-masked",
+                    { elems: { b: <b /> } },
+                  )
+                : l10n.getFragment(
+                    "landing-redesign-pricing-plans-card-free-with-bundle-feature-item-one",
+                    {
+                      vars: {
+                        data_broker_sites_total_num:
+                          CONST_ONEREP_DATA_BROKER_COUNT,
+                      },
+                      elems: { b: <b /> },
+                    },
+                  )
+              /* c8 ignore stop */
+            }
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {l10n.getFragment(
+              "landing-redesign-pricing-plans-card-free-with-bundle-feature-item-two",
+              {
+                elems: { b: <b /> },
+              },
+            )}
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {l10n.getFragment(
+              "landing-redesign-pricing-plans-card-free-with-bundle-feature-item-three",
+              {
+                elems: { b: <b /> },
+              },
+            )}
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {l10n.getFragment(
+              "landing-redesign-pricing-plans-card-free-with-bundle-feature-item-four",
+              {
+                elems: { b: <b /> },
+              },
+            )}
+          </span>
+        </>,
+        <>
+          <CheckIcon
+            alt={l10n.getString(
+              "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
+            )}
+          />
+          <span>
+            {l10n.getFragment(
+              "landing-redesign-pricing-plans-card-free-with-bundle-feature-item-five",
+              {
+                elems: { b: <b /> },
+                vars: {
+                  max_email_addresses: CONST_MAX_NUM_ADDRESSES,
+                },
+              },
+            )}
+          </span>
+        </>,
       ],
       cta: (
         <>
           <p>
-            <strong>{roundedPriceFormatter.format(0)}</strong>
+            <strong>
+              {l10n.getString(
+                "landing-redesign-pricing-plans-card-free-with-bundle-price-label",
+              )}
+            </strong>
           </p>
-          {(props.enabledFeatureFlags.includes("DisableOneRepScans") &&
-            props.eligibleForPremium) ||
-          props.scanLimitReached ? (
+          {props.scanLimitReached ? (
             <ScanLimit />
           ) : (
             <FreeScanCta
@@ -216,9 +360,11 @@ export const PricingPlanList = (props: Props & ScanLimitProp) => {
               }}
               experimentData={props.experimentData}
               ctaLabel={l10n.getString(
-                "landing-redesign-pricing-plans-card-free-cta-label",
+                "landing-redesign-pricing-plans-card-cta-label",
               )}
               showCtaOnly
+              ctaButtonVariant="secondary"
+              ctaButtonTheme="blue"
             />
           )}
         </>
@@ -232,9 +378,18 @@ export const PricingPlanList = (props: Props & ScanLimitProp) => {
         <h3>{l10n.getString("landing-redesign-pricing-plans-cards-title")}</h3>
       </VisuallyHidden>
       <span className={styles.pricingPlans}>
-        {pricingPlanData.map(
-          ({ type, label, title, subtitle, features, cta }) => (
-            <dl key={type} className={styles.pricingCard} aria-label={title}>
+        {pricingPlanData
+          .filter(
+            (pricingPlan) =>
+              !props.hideFreeCard ||
+              (props.hideFreeCard && pricingPlan.type !== "free"),
+          )
+          .map(({ type, label, title, subtitle, features, cta }) => (
+            <dl
+              key={type}
+              className={`${styles.pricingCard} ${styles[type]}`}
+              aria-label={title}
+            >
               <dt>
                 <b>{title}</b>
                 {label && (
@@ -243,19 +398,11 @@ export const PricingPlanList = (props: Props & ScanLimitProp) => {
                 <p>{subtitle}</p>
               </dt>
               {features.map((feature, featureIndex) => (
-                <dd key={`${type}-feature-${featureIndex}`}>
-                  <CheckIcon
-                    alt={l10n.getString(
-                      "landing-redesign-pricing-plans-card-plus-with-bundle-feature-item-alt-label",
-                    )}
-                  />
-                  <span>{feature}</span>
-                </dd>
+                <dd key={`${type}-feature-${featureIndex}`}>{feature}</dd>
               ))}
               <dd className={styles.pricingCardCta}>{cta}</dd>
             </dl>
-          ),
-        )}
+          ))}
       </span>
     </>
   );
