--- src/app/(proper_react)/(redesign)/(public)/LandingViewRedesign/components/PricingPlanListWithBundle.module.scss
@@ -1,204 +1,0 @@
-@use "../../../../../tokens";
-
-.pricingPlans {
-  display: flex;
-  flex-direction: column;
-  gap: tokens.$spacing-lg;
-
-  @media screen and (min-width: tokens.$screen-md) {
-    flex-wrap: wrap;
-    flex-direction: row-reverse;
-    justify-content: center;
-  }
-
-  @media screen and (min-width: tokens.$screen-lg) {
-    flex-wrap: nowrap;
-  }
-}
-
-.pricingCard {
-  box-shadow: inset 0 0 0 2px tokens.$color-grey-10;
-  border-radius: calc(tokens.$border-radius-lg + tokens.$border-radius-md);
-  display: inline-flex;
-  flex: 1;
-  flex-direction: column;
-  font: tokens.$text-body-sm;
-  max-width: tokens.$content-sm;
-  padding: tokens.$layout-sm tokens.$spacing-lg;
-  position: relative;
-
-  &:first-of-type {
-    background: linear-gradient(
-      75deg,
-      tokens.$color-violet-50 0%,
-      tokens.$color-purple-50 37.1%,
-      tokens.$color-blue-40 61.4%,
-      #0090ed 100%
-    );
-    box-shadow: none;
-
-    @media screen and (min-width: tokens.$screen-md) {
-      min-width: 100%;
-    }
-
-    @media screen and (min-width: tokens.$screen-lg) {
-      min-width: auto;
-    }
-
-    &::before {
-      background: tokens.$color-white;
-      border-radius: calc(
-        tokens.$border-radius-lg + tokens.$border-radius-md - 2px
-      );
-      content: "";
-      display: block;
-      height: calc(100% - 4px);
-      left: 2px;
-      position: absolute;
-      top: 2px;
-      width: calc(100% - 4px);
-    }
-  }
-
-  &Label {
-    color: tokens.$color-purple-70;
-    font-weight: 700;
-    left: 50%;
-    position: absolute;
-    top: tokens.$spacing-lg;
-    transform: translate(-50%);
-    white-space: nowrap;
-  }
-
-  dt,
-  dd {
-    // Ensure that the card content is not covered by the pseudo element
-    // `.pricingCard::before` that’s used for creating the border gradient of
-    // the bundle pricing card.
-    z-index: 1;
-  }
-
-  dt {
-    border-bottom: 1px solid tokens.$color-grey-10;
-    padding: tokens.$spacing-md 0;
-    text-align: center;
-
-    b {
-      font: tokens.$text-title-2xs;
-      font-weight: 700;
-      padding: tokens.$spacing-xs 0;
-    }
-
-    hr {
-      border-style: none;
-      border-top: 1px solid tokens.$color-grey-10;
-      margin: tokens.$spacing-md 0;
-      width: 100%;
-    }
-  }
-
-  dd {
-    align-items: flex-start;
-    display: flex;
-    gap: tokens.$spacing-sm;
-    padding: tokens.$spacing-sm 0;
-
-    &:first-of-type {
-      padding: tokens.$spacing-md 0 tokens.$spacing-sm;
-    }
-
-    &.pricingCardCta {
-      align-items: normal;
-      border: none;
-      display: flex;
-      flex-direction: column;
-      gap: tokens.$spacing-lg;
-      justify-content: center;
-      margin-top: auto;
-      padding: tokens.$spacing-md tokens.$spacing-xs 0;
-
-      p {
-        align-items: center;
-        display: flex;
-        gap: tokens.$spacing-xs;
-        flex-direction: column;
-      }
-
-      strong {
-        display: flex;
-        font: tokens.$text-title-2xs;
-        gap: tokens.$spacing-sm;
-        text-align: center;
-      }
-
-      s {
-        text-decoration-thickness: 0.1em;
-      }
-
-      button,
-      a {
-        width: 100%;
-      }
-    }
-
-    svg {
-      stroke: tokens.$color-purple-70;
-      flex: 1 0 24px;
-      max-width: 24px;
-      height: auto;
-    }
-  }
-
-  &.bundle dd {
-    &:first-of-type {
-      padding: tokens.$spacing-md 0 tokens.$spacing-sm;
-    }
-
-    padding: calc(tokens.$spacing-xs * 1.5) 0;
-  }
-
-  .plusNote {
-    border-bottom: 1px solid tokens.$color-grey-10;
-    align-items: center;
-    display: flex;
-    gap: tokens.$spacing-sm;
-    padding-bottom: tokens.$spacing-md;
-    width: 100%;
-  }
-
-  .bundleTitle {
-    align-items: center;
-    display: flex;
-    gap: tokens.$spacing-xs;
-    justify-content: center;
-
-    svg {
-      color: tokens.$color-black;
-      padding: calc(tokens.$spacing-xs * 0.5);
-      stroke: none;
-    }
-  }
-
-  .bundleItemLink {
-    border: 2px solid tokens.$color-grey-10;
-    border-radius: calc(tokens.$border-radius-md + tokens.$border-radius-sm);
-    color: tokens.$color-black;
-    padding: tokens.$spacing-sm;
-    text-align: center;
-    text-decoration: none;
-    width: 100%;
-
-    &:hover {
-      border: 2px solid tokens.$color-blue-50;
-      color: tokens.$color-blue-50;
-
-      svg {
-        color: tokens.$color-blue-50;
-      }
-    }
-  }
-
-  .pricingCardSavings {
-    color: tokens.$color-grey-50;
-  }
-}
