--- src/app/(proper_react)/(redesign)/(public)/LandingViewRedesign/components/PricingPlansWithBundle.module.scss
@@ -1,19 +1,0 @@
-@use "../../../../../tokens";
-
-.header,
-.content {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  max-width: tokens.$content-xl;
-}
-
-.header {
-  gap: tokens.$spacing-sm;
-  padding: 0 tokens.$layout-sm;
-  text-align: center;
-}
-
-.content {
-  gap: tokens.$layout-md;
-}
