+++ src/app/(proper_react)/(redesign)/(public)/subscription-plans/SubscriptionPlansView.tsx
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { PricingPlansWithBundle } from "../LandingViewRedesign/components/PricingPlansWithBundle";
+import { PricingPlans } from "../LandingViewRedesign/components/PricingPlans";
 import { LandingPageProps } from "../LandingViewRedesign";
 import styles from "./SubscriptionPlansView.module.scss";
 import { VisuallyHidden } from "../../../../components/server/VisuallyHidden";
@@ -13,7 +13,7 @@ export const SubscriptionPlansView = (props: LandingPageProps) => {
       <VisuallyHidden>
         <h2>{props.l10n.getString("bundle-subscription-plans-title")}</h2>
       </VisuallyHidden>
-      <PricingPlansWithBundle {...props} hideFreeCard />
+      <PricingPlans {...props} hideFreeCard />
     </main>
   );
 };
