+++ src/app/functions/server/getPremiumSubscriptionInfo.test.ts
@@ -16,6 +16,14 @@ describe("getPremiumSubscriptionInfo (SubPlat 2.0)", () => {
     expect(subscriptionUrl).toMatch(subplatUrlPattern);
   });
 
+  it("returns the yearly subscription URL", () => {
+    const subscriptionUrl = getPremiumSubscriptionUrl({
+      type: "yearly",
+      enabledFeatureFlags: [],
+    });
+    expect(subscriptionUrl).toMatch(subplatUrlPattern);
+  });
+
   it("returns the bundle subscription URL", () => {
     const subscriptionUrl = getPremiumSubscriptionUrl({
       type: "bundle",
@@ -32,6 +40,14 @@ describe("getPremiumSubscriptionInfo (SubPlat 3.0)", () => {
       enabledFeatureFlags: ["SubPlat3"],
     });
     expect(subscriptionUrl).toMatch(/\/monitorplusstage\/monthly\/landing/);
+  });
+
+  it("returns the yearly subscription URL", () => {
+    const subscriptionUrl = getPremiumSubscriptionUrl({
+      type: "yearly",
+      enabledFeatureFlags: ["SubPlat3"],
+    });
+    expect(subscriptionUrl).toMatch(/\/monitorplusstage\/yearly\/landing/);
   });
 
   it("returns the bundle subscription URL", () => {
